Revert "Modularizing some parts of the project"
