+++ .env.sample
@@ -2,4 +2,3 @@ TOKEN=
 RAW_GIT_DOC_URL=
 REGRAS_URL=
 APP_NAME=
-DEBUG=
