+++ app/bot.py
@@ -1,17 +1,89 @@
 """Modulo principal do BOT."""
-from telegram.ext import Updater
-from settings import (
-    API_KEY,
-    APP_NAME,
-    PORT
-)
-from handlers import (
-    START,
-    NEW_USER,
-    LAST_MEETUP,
-    MEETUP,
-    RULES
-)
+from decouple import config
+from telegram.ext import CommandHandler, Filters, MessageHandler, Updater
+
+from core.pugbot import PugBot
+
+API_KEY = config('TOKEN')
+APP_NAME = config('APP_NAME')
+PORT = config('PORT', default='8443', cast=int)
+
+
+def start(bot, update):
+    """Mostra um mensagem de apresentação do BOT."""
+    message = 'Olá! Sou o Bot do Python User Group - MA (PUGMA)'
+    bot.send_message(chat_id=update.message.chat_id, text=message)
+
+
+def regras(bot, update):
+    """Apresenta as regras do grupo."""
+    message = PugBot().regras()
+    bot.send_message(
+        chat_id=update.message.chat_id,
+        text=message,
+        parse_mode='Markdown'
+    )
+
+
+def hello_new_users(bot, update):
+    """Recebe um usário novo no chat do grupo."""
+    new_chat_members = update.message.new_chat_members
+
+    for member in new_chat_members:
+        is_bot = member.is_bot
+        if not is_bot:
+            message = (
+                'Olá @{}! Seja bem vindo ao Python User Group - MA (PUG-MA). '
+                'Um grupo para a galera de Python do Maranhão (ou não) que '
+                'queira interagir e ficar por dentro do que está rolando na '
+                'cena de Python aqui.'.format(member.username)
+            )
+        else:
+            message = (
+                '@{} 00101100 00100000 01101000 01100101 01101100 01101100 '
+                '01101111 00100000 01101101 01111001 '
+                '00100000 01100110 01100101 '
+                '01101100 01101100 01101111 01110111 '
+                '00100000 01101101 01100001 '
+                '01100011 01101000 01101001 01101110 '
+                '01100101 00100000 01100110 '
+                '01110010 01101001 01100101 01101110'
+                '01100100 00100001'.format(member.username)
+            )
+
+        bot.send_message(
+            chat_id=update.message.chat_id,
+            text=message
+        )
+
+        bot.send_message(
+            chat_id=update.message.chat_id,
+            text=PugBot().regras()
+        )
+
+
+def last_meetup(bot, update):
+    """Apresenta o último meetup do PUG."""
+    lastMeetup = PugBot().last_event()
+    bot.send_photo(
+        chat_id=update.message.chat_id,
+        caption=lastMeetup['text'],
+        photo=lastMeetup['photo']
+    )
+
+
+def meetup(bot, update, args):
+    """
+    Apresenta um meetup específico do PUG
+    baseado no seu número de apresentação.
+    """
+    index = int(''.join(args))
+    meetup = PugBot().event(index)
+    bot.send_photo(
+        chat_id=update.message.chat_id,
+        caption=meetup['text'],
+        photo=meetup['photo']
+    )
 
 
 def main():
@@ -19,20 +91,28 @@ def main():
     updater = Updater(token=API_KEY)
     dispatcher = updater.dispatcher
 
-    dispatcher.add_handler(START)
-    dispatcher.add_handler(NEW_USER)
-    dispatcher.add_handler(LAST_MEETUP)
-    dispatcher.add_handler(MEETUP)
-    dispatcher.add_handler(RULES)
+    start_handler = CommandHandler('start', start)
+    new_user_handler = MessageHandler(
+        Filters.status_update.new_chat_members,
+        hello_new_users
+    )
+    last_meetup_handler = CommandHandler('lastMeetup', last_meetup)
+    meetup_handler = CommandHandler('meetup', meetup, pass_args=True)
+    regras_handler = CommandHandler('regras', regras)
+
+    dispatcher.add_handler(start_handler)
+    dispatcher.add_handler(new_user_handler)
+    dispatcher.add_handler(meetup_handler)
+    dispatcher.add_handler(last_meetup_handler)
+    dispatcher.add_handler(regras_handler)
 
     updater.start_webhook(
         listen='0.0.0.0',
         port=PORT,
         url_path=API_KEY
     )
-
     updater.bot.set_webhook(f'https://{APP_NAME}.herokuapp.com/{API_KEY}')
-    #updater.start_polling()
+    # updater.start_polling()
     updater.idle()
