+++ app/core/pugbot.py
@@ -1,7 +1,7 @@
 """Modulo com principais funcionalidades do PUG-MA Bot."""
 import requests
 import json
-from settings import RAW_GIT_DOC_URL, REGRAS_URL
+from decouple import config
 
 
 class PugBot():
@@ -9,8 +9,8 @@ class PugBot():
 
     def __init__(self):
 
-        self._url = RAW_GIT_DOC_URL
-        self._regras_url = REGRAS_URL
+        self._url = config('RAW_GIT_DOC_URL')
+        self._regras_url = config('REGRAS_URL')
 
         if not self._url or not self._regras_url:
             raise Exception('Unable to load the events data, empty source URL')
